chore: extract banner-options
